infra/lambda-functions/daily_summary/requirements.txt
langchain-core
langchain_community
langchain-openai
pydantic